test: move leagcy exampe to dag
